backend/my_project/webapp/viewsets.py
from rest_framework import viewsets
from . import models
from . import serializers

class UsuarioViewset(viewsets.ModelViewSet):
    queryset = models.Usuario.objects.all()
    serializer_class = serializers.UsuarioSerializer